# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Steve  Sadr",
            "url": "https://github.com/stevesadr",
            "email": "steve.sadr@gmail.com"
        }
    },
    "paths": {
        "/v1/health/": {
            "get": {
                "description": "This handler is just hear to tell system is ready",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "System ready",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "This handler is just hear to tell post system is ready",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Post System ready",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/health/{id}": {
            "get": {
                "description": "This handler want to get id from param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Get Param",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test": {
            "get": {
                "description": "This handler Just for test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Start Testing",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/body1": {
            "post": {
                "description": "this handler hear to get response from the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing for return response",
                "parameters": [
                    {
                        "description": "person data",
                        "name": "person1",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/file1": {
            "post": {
                "description": "this handler hear to get response from the request as file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing for return response from file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file data",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/form1": {
            "post": {
                "description": "this handler hear to get response from the request as form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing for return response from form",
                "parameters": [
                    {
                        "type": "string",
                        "name": "firstName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "lastName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "mobile",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/header1": {
            "get": {
                "description": "This handler Just get token from header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing Get user token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/header2": {
            "get": {
                "description": "This handler Just get token and browser from header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing Get User Token and Browser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "browser",
                        "name": "browser",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/query1": {
            "get": {
                "description": "This handler Just for get page and tag from query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing Params query",
                "parameters": [
                    {
                        "type": "string",
                        "default": "1",
                        "description": "get page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "golang",
                        "description": "get tag",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/query2": {
            "get": {
                "description": "This handler Just for get user data from query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing User query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get user data",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/binder/uli1/{id}/{name}": {
            "get": {
                "description": "This handler get id and name from path together",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing Get Id and Name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/user/add-user": {
            "post": {
                "description": "This handler Just for test post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing Post",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/user/get-user-by-username/{username}": {
            "get": {
                "description": "This handler want to get username from the path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Test Get username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/user/{id}/account": {
            "get": {
                "description": "This handler Just for get account from the path with id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing username account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User username account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This handler Just another handler for test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Start Testing",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        },
        "/v1/test/users/{id}": {
            "get": {
                "description": "This handler want to get id from the path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Test Get id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/validations.ValidationError"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.person": {
            "type": "object",
            "required": [
                "firstName",
                "mobile"
            ],
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        },
        "helper.BaseResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "validationError": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validations.ValidationError"
                    }
                }
            }
        },
        "validations.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}